% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoneskip.R
\name{stoneskip}
\alias{stoneskip}
\title{Generate and write a stoneskip image}
\usage{
stoneskip(
  seed = 1,
  filename = NULL,
  grains = 1000,
  pixels = 5000,
  shades = 1000
)
}
\arguments{
\item{seed}{integer valued seed for the RNG state}

\item{filename}{path to image (if NULL, defaults to "stoneskip_x.png" where x is the seed)}

\item{grains}{size of the grid used for the automaton}

\item{pixels}{width and height of the final image in pixels}

\item{shades}{number of distinct colours in the image}
}
\value{
invisibly returns a matrix
}
\description{
Generate and write a stoneskip image
}
